#239: Renamed package to align with naming conventions.
